+++ Sources/SuperwallKit/Analytics/SessionEventsManager.swift
@@ -71,12 +71,6 @@ class SessionEventsManager {
 
     storage.clearCachedSessionEvents()
   }
-
-  /// This only updates the app session in the trigger sessions.
-  /// For transactions, the latest app session id is grabbed when the next transaction occurs.
-  func updateAppSession(_ appSession: AppSession) async {
-    await triggerSession.updateAppSession(to: appSession)
-  }
 }
 
 // MARK: - SessionEventsDelegate