+++ Sources/SuperwallKit/Config/Models/FeatureFlags.swift
@@ -18,10 +18,10 @@ struct FeatureFlags: Decodable {
   var enableExpressionParameters: Bool
   var enableUserIdSeed: Bool
   var disableVerboseEvents: Bool
-  var enableWebviewProcessPool: Bool
   var enableSuppressesIncrementalRendering: Bool
   var enableThrottleSchedulingPolicy: Bool
   var enableNoneSchedulingPolicy: Bool
+  var enableConfigRefresh: Bool
 
   enum CodingKeys: String, CodingKey {
     case toggles
@@ -36,13 +36,13 @@ struct FeatureFlags: Decodable {
     enablePostback = rawFeatureFlags.value(forKey: "enable_postback", default: false)
     enableUserIdSeed = rawFeatureFlags.value(forKey: "enable_userid_seed", default: false)
     disableVerboseEvents = rawFeatureFlags.value(forKey: "disable_verbose_events", default: false)
-    enableWebviewProcessPool = rawFeatureFlags.value(forKey: "enable_webview_process_pool", default: false)
     enableSuppressesIncrementalRendering = rawFeatureFlags.value(
       forKey: "enable_suppresses_incremental_rendering",
       default: false
     )
     enableThrottleSchedulingPolicy = rawFeatureFlags.value(forKey: "enable_throttle_scheduling_policy", default: false)
     enableNoneSchedulingPolicy = rawFeatureFlags.value(forKey: "enable_none_scheduling_policy", default: false)
+    enableConfigRefresh = rawFeatureFlags.value(forKey: "enable_config_refresh", default: false)
   }
 
   init(
@@ -51,20 +51,20 @@ struct FeatureFlags: Decodable {
     enableExpressionParameters: Bool,
     enableUserIdSeed: Bool,
     disableVerboseEvents: Bool,
-    enableWebviewProcessPool: Bool,
     enableSuppressesIncrementalRendering: Bool,
     enableThrottleSchedulingPolicy: Bool,
-    enableNoneSchedulingPolicy: Bool
+    enableNoneSchedulingPolicy: Bool,
+    enableConfigRefresh: Bool
   ) {
     self.enableSessionEvents = enableSessionEvents
     self.enablePostback = enablePostback
     self.enableExpressionParameters = enableExpressionParameters
     self.enableUserIdSeed = enableUserIdSeed
     self.disableVerboseEvents = disableVerboseEvents
-    self.enableWebviewProcessPool = enableWebviewProcessPool
     self.enableSuppressesIncrementalRendering = enableSuppressesIncrementalRendering
     self.enableThrottleSchedulingPolicy = enableThrottleSchedulingPolicy
     self.enableNoneSchedulingPolicy = enableNoneSchedulingPolicy
+    self.enableConfigRefresh = enableConfigRefresh
   }
 }
 
@@ -88,10 +88,10 @@ extension FeatureFlags: Stubbable {
       enableExpressionParameters: true,
       enableUserIdSeed: true,
       disableVerboseEvents: true,
-      enableWebviewProcessPool: true,
       enableSuppressesIncrementalRendering: true,
       enableThrottleSchedulingPolicy: true,
-      enableNoneSchedulingPolicy: false
+      enableNoneSchedulingPolicy: false,
+      enableConfigRefresh: true
     )
   }
 }